Create Macro1.txt
Since there is no file type for an excel macro, I uploaded it as a text file. This version auto-fills the income statement from the first 10k file. It will be relatively easy to copy this code for the balance sheet and cash flows. I still need to open the second 10k statement, check that all line items are the same, and then paste those. It would be helpful to also include dividend history long term if possible